+++ misskey-api/CHANGELOG.md
@@ -45,6 +45,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Partial support for Misskey v12.88.0
 - Support for Misskey v12.89.0
 - Support for Misskey v12.89.1 ~ v12.90.1
+- Support for Misskey v12.91.0
+   - endpoint `admin/accounts/delete`
 
 ### Changed
 ### Deprecated